// src/website/pages/holidays/model/CalenderModel.ts
export interface EventModel {
  eventId?: number;
  eventName?: string;
  eventType?: string;
  summary?: string;
  description?: string;
  location?: string;
  eventFiles?: string;
  colorCode?: string;
  organizedBy?: string;
  organizerName?: string;
  organizerMobile?: string;
  organizerEmail?: string;
  eventContactName?: string;
  eventContactMobile?: string;
  eventContactEmail?: string;
  startDate?: string;
  startTime?: string;
  endDate?: string;
  endTime?: string;
  status?: string;
}